Since the old media management application was so popular, a similar module shall be included with this release.  This means killing off the files and image applications, and creating a media application.
The problem being that this would clash with the static media folder.  Rather than break this paradigm, a new cms.lib module has been created where all non-essential extension modules live.

The idea being that all modules in lib can be trivially removed from the install, with only a few changes to the settings file.  Currently, the pages application has remained outside of lib, since it is pretty integral to the whole system.  Whether this remains the case, or whether the pages application is merged with core, remains to be seen.